move getters inside expressions module
