+++ corehq/apps/userreports/specs.py
@@ -1,7 +1,7 @@
 from jsonobject import JsonObject, StringProperty, ListProperty, BooleanProperty, DictProperty
 from jsonobject.base import DefaultProperty
 from jsonobject.exceptions import BadValueError
-from corehq.apps.userreports.getters import DictGetter, NestedDictGetter, TransformedGetter, transform_date
+from corehq.apps.userreports.expressions.getters import DictGetter, NestedDictGetter, TransformedGetter, transform_date
 from corehq.apps.userreports.operators import IN_MULTISELECT, EQUAL, OPERATORS
