+++ corehq/apps/userreports/tests/test_getters.py
@@ -1,7 +1,7 @@
 from django.test import SimpleTestCase
 from corehq.apps.userreports.exceptions import BadSpecError
 from corehq.apps.userreports.expressions.factory import ExpressionFactory
-from corehq.apps.userreports.getters import DictGetter, NestedDictGetter, TransformedGetter
+from corehq.apps.userreports.expressions.getters import DictGetter, NestedDictGetter, TransformedGetter
 
 
 class Foo(object):
